refactor(survival): move mob header files into mobs/ folder and merge shared.h into mob.h
