+++ codegen/rc4/survival/drop.c
@@ -9,7 +9,7 @@
 #include <libdl/radar.h>
 #include <libdl/string.h>
 #include "drop.h"
-#include "mob.h"
+#include "mobs/mob.h"
 #include "maputils.h"
 #include "utils.h"
 #include "game.h"
