+++ codegen/rc4/survival/mobs/mob.h
@@ -7,14 +7,14 @@
 #include <libdl/time.h>
 #include <libdl/player.h>
 #include <libdl/sound.h>
-#include "zombie.h"
-#include "executioner.h"
-#include "reactor.h"
-#include "tremor.h"
-#include "swarmer.h"
-#include "swamper.h"
-#include "reaper.h"
-#include "leviathan.h"
+#include "mobs/zombie.h"
+#include "mobs/executioner.h"
+#include "mobs/reactor.h"
+#include "mobs/tremor.h"
+#include "mobs/swarmer.h"
+#include "mobs/swamper.h"
+#include "mobs/reaper.h"
+#include "mobs/leviathan.h"
 #include "game.h"
 
 #define MOB_MAX_OTHER_TARGETS (32)
@@ -88,6 +88,16 @@ enum MobSpawnFlags
 enum MobUnreliableMsgId
 {
 	MOB_UNRELIABLE_MSG_ID_STATE_UPDATE
+};
+
+// 
+enum MOB_DO_DAMAGE_HIT_FLAGS
+{
+  MOB_DO_DAMAGE_HIT_FLAG_NONE = 0,
+  MOB_DO_DAMAGE_HIT_FLAG_HIT_TARGET = 1,
+  MOB_DO_DAMAGE_HIT_FLAG_HIT_PLAYER = 2,
+  MOB_DO_DAMAGE_HIT_FLAG_HIT_MOB = 4,
+  MOB_DO_DAMAGE_HIT_FLAG_HIT_PLAYER_THORNS = 8,
 };
 
 //
@@ -462,5 +472,44 @@ int mobHandleEvent(Moby *moby, GuberEvent *event);
 int mobCreate(int spawnParamsIdx, VECTOR position, float yaw, int spawnFromUID, int spawnFlags, struct MobConfig *config);
 void mobInitialize(void);
 void mobTick(void);
+
+int mobAmIOwner(Moby* moby);
+int mobIsFrozen(Moby* moby);
+int mobGetBehavior(Moby* moby);
+void mobResetSoundTrigger(Moby* moby);
+void mobSpawnCorn(Moby* moby, int bangle);
+int mobDoDamage(Moby* moby, float radius, float amount, int damageFlags, int friendlyFire, int jointId, int reactToThorns, int isAoE);
+int mobDoSweepDamage(Moby* moby, VECTOR from, VECTOR to, float step, float radius, float amount, int damageFlags, int friendlyFire, int reactToThorns, int isAoE);
+int mobDoDamageTryHit(Moby* moby, Moby* hitMoby, VECTOR jointPosition, int isAoE, float hitRadius, int damageFlags, float amount);
+void mobSetAction(Moby* moby, int action);
+void mobTransAnimLerp(Moby* moby, int animId, int lerpFrames, float startOff);
+void mobTransAnim(Moby* moby, int animId, float startOff);
+int mobHasVelocity(struct MobPVar* pvars);
+float mobGetCurrentMoveSpeed(Moby* moby);
+void mobGetKnockbackVelocity(Moby* moby, VECTOR out);
+void mobGetTargetCenter(Moby* target, VECTOR out);
+int mobCanSeeMoby(Moby* moby, Moby* canSeeMoby);
+void mobStand(Moby* moby);
+void mobResetMoveStep(Moby* moby);
+int mobMoveCheck(Moby* moby, VECTOR outputPos, VECTOR from, VECTOR to);
+void mobMove(Moby* moby);
+void mobMoveTowards(Moby* moby, VECTOR targetPosition, float speed, float turnSpeed, float acceleration, float curveNearTargetDir);
+void mobJumpTowards(Moby* moby, VECTOR targetPosition);
+int mobHitWallShouldJump(Moby* moby, float maxSlope);
+float mobTurnTowards(Moby* moby, VECTOR towards, float turnSpeed);
+float mobTurnTowardsPredictive(Moby* moby, Moby* target, float turnSpeed, float predictFactor);
+void mobGetVelocityToTargetWithDirection(Moby* moby, VECTOR velocity, VECTOR from, VECTOR to, float yaw, float speed, float acceleration);
+void mobGetVelocityToTarget(Moby* moby, VECTOR velocity, VECTOR from, VECTOR to, float speed, float acceleration);
+void mobGetVelocityToTargetSimple(Moby* moby, VECTOR velocity, VECTOR from, VECTOR to, float speed, float acceleration);
+void mobPostDrawQuad(Moby* moby, float scale, u32 color, int jointId);
+void mobOnStateUpdate(Moby* moby, struct MobStateUpdateEventArgs* e);
+void mobPreUpdate(Moby* moby);
+int mobIsProjectileComing(Moby* moby);
+float mobGetCurrentWalkAngle(Moby* moby);
+float mobGetScaleMultiplier(Moby* moby);
+Moby* mobGetNextTarget(Moby* moby, float keepCurrentTargetFactor);
+void mobDefaultPreUpdate(Moby *moby);
+int mobDefaultOnLocalDamage(Moby *moby, struct MobLocalDamageEventArgs *e);
+void mobHandleFlinch(Moby *moby, struct MobDamageEventArgs *e, int canFlinch, int isShock, float probability, float powerFactor, int flinchAction, int bigFlinchAction);
 
 #endif // SURVIVAL_MOB_H
