+++ codegen/rc4/survival/mobs/reactor.h
