+++ codegen/rc4/survival/mobs/swamper.c
@@ -9,11 +9,10 @@
 #include <libdl/color.h>
 
 #include "game.h"
-#include "mob.h"
+#include "mobs/mob.h"
 #include "pathfind.h"
 #include "utils.h"
 #include "maputils.h"
-#include "shared.h"
 
 void swamperPreUpdate(Moby *moby);
 void swamperPostUpdate(Moby *moby);
