+++ codegen/rc4/survival/mobs/swamper.h
