+++ codegen/rc4/survival/mobs/swarmer.c
@@ -9,11 +9,10 @@
 #include <libdl/color.h>
 
 #include "game.h"
-#include "mob.h"
+#include "mobs/mob.h"
 #include "pathfind.h"
 #include "utils.h"
 #include "maputils.h"
-#include "shared.h"
 
 void swarmerPreUpdate(Moby *moby);
 void swarmerPostUpdate(Moby *moby);
