+++ codegen/rc4/survival/mobs/swarmer.h
