+++ codegen/rc4/survival/mobs/trailshot.c
@@ -8,11 +8,10 @@
 #include <libdl/color.h>
 
 #include "game.h"
-#include "trailshot.h"
-#include "mob.h"
+#include "mobs/trailshot.h"
+#include "mobs/mob.h"
 #include "utils.h"
 #include "maputils.h"
-#include "shared.h"
 
 VECTOR trailshotCenterOffset = {0, 0, 1, 0};
 int trailshotGlobalIndex = 0;
