+++ codegen/rc4/survival/mobs/tremor.c
@@ -8,10 +8,9 @@
 #include <libdl/color.h>
 
 #include "game.h"
-#include "mob.h"
+#include "mobs/mob.h"
 #include "utils.h"
 #include "maputils.h"
-#include "shared.h"
 
 void tremorPreUpdate(Moby *moby);
 void tremorPostUpdate(Moby *moby);
